test: DeckGeneratorTest DeckTest로 위임
